requirements.txt
open3d
numpy
matplotlib
libigl
pymeshlab
tkinter



